Added -l parm to split output based on GMap layers
Realized that some GMap's have layers that are useful when planning a trip, but you might not want to see all the layers when you're using the map in OSMAnd.  If you used layers in the GMap preserving that structure probably makes sense in OSMAnd as well.  Using OSMAnd my places you can select the appropriate folder and turn off it's tracks and waypts with one operation instead of selecting all the items individually.